+++ app/src/main/java/ke/co/xently/shopping/datasource/remote/CacheControl.kt
@@ -1,4 +1,4 @@
-package ke.co.xently.shopping.remotedatasource
+package ke.co.xently.shopping.datasource.remote
 
 sealed class CacheControl(private val name: String) {
     override fun toString(): String {
