+++ app/src/main/java/ke/co/xently/shopping/datasource/remote/HttpException.kt
@@ -1,4 +1,4 @@
-package ke.co.xently.shopping.remotedatasource
+package ke.co.xently.shopping.datasource.remote
 
 open class HttpException(
     val detail: Any? = null,
