+++ app/src/main/java/ke/co/xently/shopping/datasource/remote/exceptions/WebsocketConnectionFailedException.kt
@@ -1,3 +1,3 @@
-package ke.co.xently.shopping.remotedatasource.exceptions
+package ke.co.xently.shopping.datasource.remote.exceptions
 
 class WebsocketConnectionFailedException : WebsocketException("Failed to establish a connection.")
