+++ app/src/main/java/ke/co/xently/shopping/features/attributes/datasources/remoteservices/AttributeService.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.attributes.datasources.remoteservices
 
 import ke.co.xently.shopping.features.attributes.models.Attribute
 import ke.co.xently.shopping.features.core.models.RemoteSearchResponse
-import ke.co.xently.shopping.remotedatasource.CacheControl
+import ke.co.xently.shopping.datasource.remote.CacheControl
 import retrofit2.Response
 import retrofit2.http.*
