+++ app/src/main/java/ke/co/xently/shopping/features/compareproducts/datasources/RemoteCompareProductDataSource.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.compareproducts.datasources
 
 import ke.co.xently.shopping.features.compareproducts.datasources.remoteservices.CompareProductService
 import ke.co.xently.shopping.features.compareproducts.models.CompareProduct
-import ke.co.xently.shopping.remotedatasource.SendHttpRequest
+import ke.co.xently.shopping.datasource.remote.SendHttpRequest
 import javax.inject.Inject
 import javax.inject.Singleton
