+++ app/src/main/java/ke/co/xently/shopping/features/measurementunit/datasources/RemoteMeasurementUnitDataSource.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.measurementunit.datasources
 
 import ke.co.xently.shopping.features.measurementunit.datasources.remoteservices.MeasurementUnitService
 import ke.co.xently.shopping.features.measurementunit.models.MeasurementUnit
-import ke.co.xently.shopping.remotedatasource.SendHttpRequest
+import ke.co.xently.shopping.datasource.remote.SendHttpRequest
 import javax.inject.Inject
 import javax.inject.Singleton
