+++ app/src/main/java/ke/co/xently/shopping/features/products/datasources/RemoteProductDataSource.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.products.datasources
 
 import ke.co.xently.shopping.features.products.datasources.remoteservices.ProductService
 import ke.co.xently.shopping.features.products.models.Product
-import ke.co.xently.shopping.remotedatasource.SendHttpRequest
+import ke.co.xently.shopping.datasource.remote.SendHttpRequest
 import javax.inject.Inject
 import javax.inject.Singleton
