+++ app/src/main/java/ke/co/xently/shopping/features/recommendations/datasources/RemoteRecommendationDataSource.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.recommendations.datasources
 
 import ke.co.xently.shopping.features.recommendations.datasources.remoteservices.RecommendationService
 import ke.co.xently.shopping.features.recommendations.models.Recommendation
-import ke.co.xently.shopping.remotedatasource.SendHttpRequest
+import ke.co.xently.shopping.datasource.remote.SendHttpRequest
 import javax.inject.Inject
 import javax.inject.Singleton
