+++ app/src/main/java/ke/co/xently/shopping/features/store/datasources/RemoteStoreDataSource.kt
@@ -2,7 +2,7 @@ package ke.co.xently.shopping.features.store.datasources
 
 import ke.co.xently.shopping.features.store.datasources.remoteservices.StoreService
 import ke.co.xently.shopping.features.store.models.Store
-import ke.co.xently.shopping.remotedatasource.SendHttpRequest
+import ke.co.xently.shopping.datasource.remote.SendHttpRequest
 import javax.inject.Inject
 import javax.inject.Singleton
